apps/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy>=2.0.30
alembic>=1.13.2
psycopg2-binary==2.9.9; sys_platform != 'darwin'
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.2.1
boto3==1.34.69
redis==5.0.3
rq==1.16.2
music21
email-validator>=2.1.0
pandas>=2.0.0
openpyxl>=3.1.0
